added capability to up 10000mhz

+++ src/DemodulateSignal.java
@@ -28,7 +28,7 @@ public class DemodulateSignal implements IEvaluableEnTiempo {
     }
 
     @Override
-    public int unidadOptima() {
+    public double unidadOptima() {
        return parent.unidadOptima();
     }
